Rename package to dukegg
