+++ src/main/java/commands/ByeCommand.java
@@ -1,8 +1,8 @@
 package commands;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import dukeegg.Storage;
+import dukeegg.TaskList;
+import dukeegg.Ui;
 
 /**
  * Terminates the chatbot program after execution, and saves the current tasks to some local storage.
