+++ src/main/java/commands/EventCommand.java
@@ -3,9 +3,9 @@ package commands;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
 
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import dukeegg.Storage;
+import dukeegg.TaskList;
+import dukeegg.Ui;
 import exceptions.DukeException;
 import exceptions.EmptyDateTimeException;
 import exceptions.EmptyDescriptionException;
