+++ src/main/java/dukeegg/DialogBox.java
@@ -1,4 +1,4 @@
-package duke;
+package dukeegg;
 
 import java.io.IOException;
 import java.util.Collections;
