+++ src/main/java/dukeegg/Main.java
@@ -1,4 +1,4 @@
-package duke;
+package dukeegg;
 
 import java.io.IOException;
