+++ src/main/java/dukeegg/Storage.java
@@ -1,4 +1,4 @@
-package duke;
+package dukeegg;
 
 import java.io.File;
 import java.io.FileWriter;
