+++ src/main/java/dukeegg/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package dukeegg;
 
 import task.Task;
