+++ src/test/java/dukeegg/TaskListTest.java
@@ -1,4 +1,4 @@
-package duke;
+package dukeegg;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
